refactor: package name 변경 study > stage1
