+++ src/test/java/stage1/SetTest.java
@@ -1,4 +1,4 @@
-package study;
+package stage1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
