+++ src/test/java/stage1/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package stage1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
